tests/integration/requirements.txt
boto3
pytest-asyncio
asyncio
